Extract the isSneaking check into a new util-class and move some other interaction related utilities there.
