+++ src/main/java/appeng/block/misc/CondenserBlock.java
@@ -34,6 +34,7 @@ import appeng.container.ContainerLocator;
 import appeng.container.ContainerOpener;
 import appeng.container.implementations.CondenserContainer;
 import appeng.tile.misc.CondenserTileEntity;
+import appeng.util.InteractionUtil;
 import appeng.util.Platform;
 
 public class CondenserBlock extends AEBaseTileBlock<CondenserTileEntity> {
@@ -45,13 +46,13 @@ public class CondenserBlock extends AEBaseTileBlock<CondenserTileEntity> {
     @Override
     public ActionResultType onActivated(final World w, final BlockPos pos, final PlayerEntity player, final Hand hand,
             final @Nullable ItemStack heldItem, final BlockRayTraceResult hit) {
-        if (Platform.isEntityHoldingShift(player)) {
+        if (InteractionUtil.isInAlternateUseMode(player)) {
             return ActionResultType.PASS;
         }
 
         if (Platform.isServer()) {
             final CondenserTileEntity tc = this.getTileEntity(w, pos);
-            if (tc != null && !Platform.isEntityHoldingShift(player)) {
+            if (tc != null && !InteractionUtil.isInAlternateUseMode(player)) {
                 ContainerOpener.openContainer(CondenserContainer.TYPE, player,
                         ContainerLocator.forTileEntitySide(tc, hit.getFace()));
                 return ActionResultType.SUCCESS;
