+++ src/main/java/appeng/debug/ReplicatorCardItem.java
@@ -46,7 +46,7 @@ import appeng.api.networking.spatial.ISpatialCache;
 import appeng.api.util.AEPartLocation;
 import appeng.api.util.DimensionalCoord;
 import appeng.items.AEBaseItem;
-import appeng.util.Platform;
+import appeng.util.InteractionUtil;
 
 public class ReplicatorCardItem extends AEBaseItem {
 
@@ -95,7 +95,7 @@ public class ReplicatorCardItem extends AEBaseItem {
         int y = pos.getY();
         int z = pos.getZ();
 
-        if (Platform.isEntityHoldingShift(player)) {
+        if (InteractionUtil.isInAlternateUseMode(player)) {
             if (world.getTileEntity(pos) instanceof IGridHost) {
                 final CompoundNBT tag = player.getHeldItem(hand).getOrCreateTag();
                 tag.putInt("x", x);
