+++ src/main/java/appeng/items/materials/MaterialItem.java
@@ -45,8 +45,8 @@ import appeng.api.implementations.tiles.ISegmentedInventory;
 import appeng.api.parts.IPartHost;
 import appeng.api.parts.SelectedPart;
 import appeng.items.AEBaseItem;
+import appeng.util.InteractionUtil;
 import appeng.util.InventoryAdaptor;
-import appeng.util.Platform;
 import appeng.util.inv.AdaptorItemHandler;
 
 public final class MaterialItem extends AEBaseItem implements IStorageComponent, IUpgradeModule {
@@ -108,7 +108,7 @@ public final class MaterialItem extends AEBaseItem implements IStorageComponent,
     public ActionResultType onItemUseFirst(ItemStack stack, ItemUseContext context) {
         PlayerEntity player = context.getPlayer();
         Hand hand = context.getHand();
-        if (Platform.isEntityHoldingShift(player)) {
+        if (InteractionUtil.isInAlternateUseMode(player)) {
             final TileEntity te = context.getWorld().getTileEntity(context.getPos());
             IItemHandler upgrades = null;
