+++ src/main/java/appeng/items/storage/AbstractStorageCell.java
@@ -49,6 +49,7 @@ import appeng.items.AEBaseItem;
 import appeng.items.contents.CellConfig;
 import appeng.items.contents.CellUpgrades;
 import appeng.items.materials.MaterialType;
+import appeng.util.InteractionUtil;
 import appeng.util.InventoryAdaptor;
 import appeng.util.Platform;
 
@@ -140,11 +141,11 @@ public abstract class AbstractStorageCell<T extends IAEStack<T>> extends AEBaseI
     }
 
     private boolean disassembleDrive(final ItemStack stack, final World world, final PlayerEntity player) {
-        if (Platform.isClient()) {
-            return false;
-        }
+        if (InteractionUtil.isInAlternateUseMode(player)) {
+            if (Platform.isClient()) {
+                return false;
+            }
 
-        if (Platform.isEntityHoldingShift(player)) {
             final PlayerInventory playerInventory = player.inventory;
             final IMEInventoryHandler inv = Api.instance().registries().cell().getCellInventory(stack, null,
                     this.getChannel());
