+++ src/main/java/appeng/parts/PartPlacement.java
@@ -63,6 +63,7 @@ import appeng.core.sync.network.NetworkHandler;
 import appeng.core.sync.packets.ClickPacket;
 import appeng.core.sync.packets.PartPlacementPacket;
 import appeng.facade.IFacadeItem;
+import appeng.util.InteractionUtil;
 import appeng.util.LookDirection;
 import appeng.util.Platform;
 
@@ -79,13 +80,14 @@ public class PartPlacement {
         }
 
         // FIXME: This was changed alot.
-        final LookDirection dir = Platform.getPlayerRay(player);
+        final LookDirection dir = InteractionUtil.getPlayerRay(player);
         RayTraceContext rtc = new RayTraceContext(dir.getA(), dir.getB(), RayTraceContext.BlockMode.OUTLINE,
                 RayTraceContext.FluidMode.NONE, player);
         final BlockRayTraceResult mop = world.rayTraceBlocks(rtc);
         BlockItemUseContext useContext = new BlockItemUseContext(new ItemUseContext(player, hand, mop));
 
-        if (!held.isEmpty() && Platform.isWrench(player, held, pos) && Platform.isEntityHoldingShift(player)) {
+        if (!held.isEmpty() && InteractionUtil.isWrench(player, held, pos)
+                && InteractionUtil.isInAlternateUseMode(player)) {
             if (!Platform.hasPermissions(new DimensionalCoord(world, pos), player)) {
                 return ActionResultType.FAIL;
             }
@@ -179,7 +181,7 @@ public class PartPlacement {
         }
 
         if (held.isEmpty()) {
-            if (host != null && Platform.isEntityHoldingShift(player) && world.isAirBlock(pos)) {
+            if (host != null && InteractionUtil.isInAlternateUseMode(player) && world.isAirBlock(pos)) {
                 if (mop.getType() == RayTraceResult.Type.BLOCK) {
                     Vector3d hitVec = mop.getHitVec().add(-mop.getPos().getX(), -mop.getPos().getY(),
                             -mop.getPos().getZ());
@@ -286,7 +288,8 @@ public class PartPlacement {
                         mop.getHitVec().add(-mop.getPos().getX(), -mop.getPos().getY(), -mop.getPos().getZ()));
 
                 if (sp.part != null) {
-                    if (!Platform.isEntityHoldingShift(player) && sp.part.onActivate(player, hand, mop.getHitVec())) {
+                    if (!InteractionUtil.isInAlternateUseMode(player)
+                            && sp.part.onActivate(player, hand, mop.getHitVec())) {
                         return ActionResultType.FAIL;
                     }
                 }
@@ -323,7 +326,7 @@ public class PartPlacement {
 
     private static float getEyeOffset(final PlayerEntity p) {
         if (p.world.isRemote) {
-            return Platform.getEyeOffset(p);
+            return InteractionUtil.getEyeOffset(p);
         }
 
         return getEyeHeight();
@@ -361,7 +364,7 @@ public class PartPlacement {
         if (event instanceof PlayerInteractEvent.RightClickEmpty && event.getPlayer().world.isRemote) {
             // re-check to see if this event was already channeled, cause these two events
             // are really stupid...
-            final RayTraceResult mop = Platform.rayTrace(event.getPlayer(), true, false);
+            final RayTraceResult mop = InteractionUtil.rayTrace(event.getPlayer(), true, false);
             final Minecraft mc = Minecraft.getInstance();
 
             final float f = 1.0F;
@@ -383,7 +386,7 @@ public class PartPlacement {
                 boolean supportedItem = items.memoryCard().isSameAs(held);
                 supportedItem |= items.colorApplicator().isSameAs(held);
 
-                if (Platform.isEntityHoldingShift(event.getPlayer()) && !held.isEmpty() && supportedItem) {
+                if (InteractionUtil.isInAlternateUseMode(event.getPlayer()) && !held.isEmpty() && supportedItem) {
                     NetworkHandler.instance().sendToServer(new ClickPacket(event.getHand()));
                 }
             }
